Fix marker identity, removal, and trackline pause semantics
Migrates marker operations (remove/toggle) from time-based lookups to robust, deterministic u64 IDs. Fixes trackline pausing to actually respect the `pause_at` flag of individual markers. Enhances the timeline UI to allow middle-clicking to toggle marker pause state with clear visual feedback.

+++ crates/vorce-ui/src/editors/timeline_v2/types.rs
@@ -10,7 +10,7 @@ pub enum TimelineAction {
     SelectModule(ModuleId),
     AddMarker(f32),
     RemoveMarker(u64),
-    ToggleMarkerPause(f32),
+    ToggleMarkerPause(u64),
     JumpNextMarker,
     JumpPrevMarker,
 }
